+++ src/pimApi.spec.js
@@ -1,3 +1,4 @@
+
 import expect from "must";
 import express from "express";
 import {getAllTables, getTablesByNames, getCompleteTable} from "./pimApi";
@@ -31,6 +32,10 @@ describe("pimApi", () => {
         } else if (req.url === "/tables/3/rows?offset=2&limit=2") {
           res.sendFile(`${__dirname}/__tests__/pimFixture-3-rows2.json`);
         } else if (req.url === "/tables/3/rows?offset=4&limit=2") {
+          res.sendFile(`${__dirname}/__tests__/pimFixture-3-rows3.json`);
+        } else if (req.url === "/tables/3/rows?offset=0&limit=2&archived=true") {
+          res.sendFile(`${__dirname}/__tests__/pimFixture-3-rows3.json`);
+        } else if (req.url === "/tables/3/rows?offset=0&limit=2&archived=false") {
           res.sendFile(`${__dirname}/__tests__/pimFixture-3-rows3.json`);
         } else {
           res.end("error");
@@ -172,6 +177,45 @@ describe("pimApi", () => {
         expect(result).to.have.property("rows");
         expect(result.rows).to.be.an.array();
         expect(result.rows.length).to.equal(5);
+      });
+    });
+
+    it("should construct the correct URL for the rows API call without setting the archived parameter", () => {
+      return getCompleteTable({
+        pimUrl: SERVER_URL
+      }, 3, 2).then(result => {
+        expect(calledUrls.some(elem => /\/tables\/3\/rows\?offset=0&limit=2/.test(elem))).to.be.true();
+
+        expect(result).to.have.property("id");
+        expect(result.name).to.equal("thirdTestTable");
+        expect(result).to.have.property("rows");
+        expect(result.rows).to.be.an.array();
+      });
+    });
+
+    it("should construct the correct URL for the rows API call with archived parameter set to true", () => {
+      return getCompleteTable({
+        pimUrl: SERVER_URL
+      }, 3, 2, true).then(result => {
+        expect(calledUrls.some(elem => /\/tables\/3\/rows\?offset=0&limit=2&archived=true/.test(elem))).to.be.true();
+
+        expect(result).to.have.property("id");
+        expect(result.name).to.equal("thirdTestTable");
+        expect(result).to.have.property("rows");
+        expect(result.rows).to.be.an.array();
+      });
+    });
+
+    it("should construct the correct URL for the rows API call with archived parameter set to false", () => {
+      return getCompleteTable({
+        pimUrl: SERVER_URL
+      }, 3, 2, false).then(result => {
+        expect(calledUrls.some(elem => /\/tables\/3\/rows\?offset=0&limit=2&archived=false/.test(elem))).to.be.true();
+
+        expect(result).to.have.property("id");
+        expect(result.name).to.equal("thirdTestTable");
+        expect(result).to.have.property("rows");
+        expect(result.rows).to.be.an.array();
       });
     });